pyproject.toml
[tool.poetry]
name = "llm-project-summarizer"
version = "0.0.2"
description = "Generate LLM-friendly summaries of Python and Go codebases"
authors = ["Yossry"]
homepage = "https://github.com/MuhamadYossry/llm-project-summarizer"
repository = "https://github.com/MuhamadYossry/llm-project-summarizer"
keywords = ["llm"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
click = "^8.1.8"
pyyaml = "^6.0.2"
tqdm = "^4.67.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.0"
pylint = "^3.3.2"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
